Fixed all warnings with recent Pythons and when compiling with C++17

Fixed Windows platform CI build

TODO: diagnose fix red tests on windows platform (Debug build linking issue and Release build segfault)

https://github.com/reefactor/cppy3/actions/runs/10724591615/job/29740578852?pr=6

+++ cppy3/cppy3_numpy.hpp
@@ -103,7 +103,7 @@ namespace cppy3
      * @param n2 - cols size
      * @param fillZeros - initialize allocated array with zeros
      */
-    void create(int n1, int n2, bool fillZeros = SLOWER_AND_CLEARNER)
+    void create(size_t n1, size_t n2, bool fillZeros = SLOWER_AND_CLEARNER)
     {
 
       decref();
